+++ docs/constant/members.ts
@@ -1,0 +1,164 @@
+export const MEMBERS = [
+  {
+    name: { en: "Ryota Nakazono", ja: "Ryota Nakazono" },
+    description: {
+      en: "Front-end engineer, developing web applications.",
+      ja: "フロントエンドエンジニア。Webアプリケーションの開発を行っている。",
+    },
+    github: {
+      id: "zonoryo17",
+      icon: "https://avatars.githubusercontent.com/u/88890984?v=4",
+      url: "https://github.com/zonoryo17",
+    },
+    x: { url: "https://x.com/zonoryo03" },
+    discord: { id: "968897190783246416" },
+  },
+  {
+    name: { en: "Hayato Nakamura", ja: "Hayato Nakamura" },
+    description: {
+      en: "A college student who enjoys teaching programming to middle and high school students.",
+      ja: "中高生にプログラミングを教えることを楽しんでいる大学生。",
+    },
+    github: {
+      id: "Hayato1031",
+      icon: "https://avatars.githubusercontent.com/u/105223982?v=4",
+      url: "https://github.com/Hayato1031",
+    },
+    x: { url: "https://x.com/lit_myura" },
+    discord: { id: "663701704990130177" },
+  },
+  {
+    name: { en: "Kurakake Kaito", ja: "Kurakake Kaito" },
+    description: {
+      en: "A student engineer who enjoys front-end and back-end development.",
+      ja: "フロントエンドとバックエンドを楽しむ学生エンジニア。",
+    },
+    github: {
+      id: "kurakke",
+      icon: "https://avatars.githubusercontent.com/u/84614038?v=4",
+      url: "https://github.com/kurakke",
+    },
+    x: { url: "https://x.com/kurakke" },
+    discord: { id: "1009394043974856704" },
+  },
+  {
+    name: { en: "Haruki Maeda", ja: "Haruki Maeda" },
+    description: { en: "Engineer based in tokyo.", ja: "東京のエンジニア。" },
+    github: {
+      id: "Haru-0001",
+      icon: "https://avatars.githubusercontent.com/u/166319111?v=4",
+      url: "https://github.com/Haru-0001",
+    },
+    discord: { id: "1244260714919497809" },
+  },
+  {
+    name: { en: "Keiju Fujiwara", ja: "Keiju Fujiwara" },
+    description: {
+      en: "Back-end engineer currently studying front-end engineering.",
+      ja: "フロント勉強中の普段はバックエンドエンジニア。",
+    },
+    github: {
+      id: "fffk10",
+      icon: "https://avatars.githubusercontent.com/u/77735745?v=4",
+      url: "https://github.com/fffk10",
+    },
+    x: { url: "https://x.com/w_q_r_q" },
+    discord: { id: "385445321263677440" },
+  },
+  {
+    name: { en: "Tomoya Matsubara", ja: "Tomoya Matsubara" },
+    description: {
+      en: "From Web development to machine learning.",
+      ja: "Web開発から機械学習まで何でもやるエンジニア。",
+    },
+    github: {
+      id: "Tomoya-Matsubara",
+      icon: "https://avatars.githubusercontent.com/u/92769722?v=4",
+      url: "https://github.com/Tomoya-Matsubara",
+    },
+    discord: { id: "1176786237944766495" },
+  },
+  {
+    name: { en: "Hiro Nakanishi", ja: "Hiro Nakanishi" },
+    description: {
+      en: "Security and machine learning engineer.",
+      ja: "セキュリティと機械学習エンジニア。",
+    },
+    github: {
+      id: "quantumshiro",
+      icon: "https://avatars.githubusercontent.com/u/75029815?v=4",
+      url: "https://github.com/quantumshiro",
+    },
+    x: { url: "https://x.com/Curiosi46542428" },
+    discord: { id: "528930667094867978" },
+  },
+  {
+    name: { en: "Ryota Suzuki", ja: "Ryota Suzuki" },
+    description: {
+      en: "Engineer living in Mie Prefecture, Japan.",
+      ja: "三重県在住エンジニア。",
+    },
+    github: {
+      id: "suzukisan22",
+      icon: "https://avatars.githubusercontent.com/u/12987013?v=4",
+      url: "https://github.com/suzukisan22",
+    },
+    x: { url: "https://x.com/itsmeryota" },
+    discord: { id: "611441723712864256" },
+  },
+  {
+    name: { en: "Haru", ja: "Haru" },
+    description: {
+      en: "University students studying programming and design.",
+      ja: "プログラミングとデザインを学んでいる大学生。",
+    },
+    github: {
+      id: "haru-036",
+      icon: "https://avatars.githubusercontent.com/u/105186692?v=4",
+      url: "https://github.com/haru-036",
+    },
+    x: { url: "https://x.com/haru_u_de" },
+    discord: { id: "1063410330153259078" },
+  },
+  {
+    name: { en: "Kaito Suzuki", ja: "Kaito Suzuki" },
+    description: {
+      en: "A student engineer who is studying front-end and back-end.",
+      ja: "フロントエンドとバックエンドを勉強している学生エンジニア。",
+    },
+    github: {
+      id: "kaitobq",
+      icon: "https://avatars.githubusercontent.com/u/157810570?v=4",
+      url: "https://github.com/kaitobq",
+    },
+    x: { url: "https://x.com/kaitobq" },
+    discord: { id: "349526549474705408" },
+  },
+  {
+    name: { en: "Shuta Yanagawa", ja: "Shuta Yanagaawa" },
+    description: {
+      en: "Web front-end engineer. a NIT Student.",
+      ja: "Webフロントエンドエンジニア。高専生。",
+    },
+    github: {
+      id: "Shunii85",
+      icon: "https://avatars.githubusercontent.com/u/137262383?v=4",
+      url: "https://github.com/Shunii85",
+    },
+    x: { url: "https://x.com/shunii_hack" },
+    discord: { id: "1098389261738328084" },
+  },
+  {
+    name: { en: "Motohiro Yamagishi", ja: "Motohiro Yamagishi" },
+    description: {
+      en: "Front-end engineer, a little bit of back-end.",
+      ja: "フロントエンドエンジニア、少しだけバックエンド。",
+    },
+    github: {
+      id: "mokun632",
+      icon: "https://avatars.githubusercontent.com/mokun632",
+      url: "https://github.com/mokun632",
+    },
+    discord: { id: "1221042700304912437" },
+  },
+]