+++ packages/components/tooltip/tests/tooltip.test.tsx
@@ -1,7 +1,7 @@
 import { a11y, fireEvent, render, screen, waitFor } from "@yamada-ui/test"
 import { Tooltip } from "../src/tooltip"
 
-describe("<Tooltip/>", () => {
+describe("<Tooltip />", () => {
   test("should pass a11y test", async () => {
     await a11y(
       <Tooltip label="Tooltip hover">
@@ -77,4 +77,67 @@ describe("<Tooltip/>", () => {
     const tooltip = screen.queryByRole("tooltip")
     expect(tooltip).toBeNull()
   })
+
+  test("should not render label text when `Escape` pressed", async () => {
+    render(
+      <Tooltip label="Tooltip hover">
+        <span>Hover</span>
+      </Tooltip>,
+    )
+
+    fireEvent.pointerEnter(screen.getByText("Hover"))
+
+    await waitFor(() => {
+      expect(screen.getAllByText("Tooltip hover")[1]).toBeInTheDocument()
+    })
+
+    fireEvent.keyDown(screen.getByText("Hover"), { key: "Escape" })
+
+    await waitFor(() => {
+      expect(screen.queryByRole("tooltip")).toBeNull()
+    })
+  })
+
+  test.each([
+    {
+      closeOnMouseDown: false,
+      closeOnPointerDown: true,
+    },
+    {
+      closeOnMouseDown: true,
+      closeOnPointerDown: false,
+    },
+    {
+      closeOnMouseDown: true,
+      closeOnPointerDown: true,
+    },
+  ])(
+    "Tooltip should be hidden (when `closeOnMouseDown` is $closeOnMouseDown and `closeOnPointerDown` is $closeOnPointerDown)",
+    async ({ closeOnMouseDown, closeOnPointerDown }) => {
+      const { user } = render(
+        <Tooltip
+          label="Tooltip hover"
+          closeOnMouseDown={closeOnMouseDown}
+          closeOnPointerDown={closeOnPointerDown}
+        >
+          <span>Hover</span>
+        </Tooltip>,
+      )
+
+      const tooltipTriggerElement = await screen.findByText("Hover")
+      await user.hover(tooltipTriggerElement)
+
+      const tooltip = await screen.findByRole("tooltip", {
+        name: /Tooltip hover/i,
+      })
+      expect(tooltip).toBeVisible()
+
+      await user.pointer({ target: tooltipTriggerElement, keys: "[MouseLeft]" })
+      await waitFor(() => {
+        expect(
+          screen.queryByRole("tooltip", { name: /Tooltip hover/i }),
+        ).toBeNull()
+      })
+    },
+  )
 })