+++ packages/utils/src/index.types.ts
@@ -21,7 +21,7 @@ export type Dict<T = any> = Record<string, T>
 
 export type ObjectLiteral = {}
 
-export type StringLiteral = string & ObjectLiteral
+export type StringLiteral = string & {}
 
 export type Union<T> = T | StringLiteral